Working on adding progress state to the program through the IDialogService. Progress most likely need to go through the ViewModel class, currently it's being accessed through the service classes and it's not working right (UI is not updated). This is still .Net 6 - going to try and upgrade the project to .Net 8.

+++ src/CoreUI/Services/CountryService.cs
@@ -73,7 +73,7 @@ namespace TradeSharp.CoreUI.Services
 
     public Task<Country> CopyAsync(Country item) => throw new NotImplementedException();
     public Task<ImportReplaceResult> ImportAsync(ImportSettings importSettings) => throw new NotImplementedException();
-    public Task<int> ExportAsync(string filename) => throw new NotImplementedException();
+    public Task<long> ExportAsync(string filename) => throw new NotImplementedException();
 
     //properties
     public event EventHandler<Country>? SelectedItemChanged;
