+++ src/CoreUI/Services/SessionService.cs
@@ -96,6 +96,6 @@ namespace TradeSharp.CoreUI.Services
     }
 
     public Task<ImportReplaceResult> ImportAsync(ImportSettings importSettings) => throw new NotImplementedException();
-    public Task<int> ExportAsync(string filename) => throw new NotImplementedException();
+    public Task<long> ExportAsync(string filename) => throw new NotImplementedException();
   }
 }
